Ensure normalise doesn't care whether var is bound
Normalise would previously fail if it encountered an unbound variable,
but this wasn't very useful if e.g., we wanted to synthesise the type
of a signature containing level variables. Not 100% sure this is the
way we definitely want to go, but I don't think it will bug
out... Hopefully.